fix: update IPFS websocket reconnect logic (#785)
Previously, the `_maintainConnection()` function would disconnect and reconnect
to all bootstrap nodes every 10 seconds. One of the issues we had was that the
call to `node.swarm.connect()` would occasionally throw an error that we didn't
catch, so the `_maintainConnection()` function would die. This commit adds a
try/catch around the error.

Additionally, this commit checks the nodes that we're already connected to, and
only attempts to reconnect to any bootstrap nodes that we disconnected from at
some point.

Co-authored-by: David Grisham <grish@nixmail.dev>